+++ src/components/Modal/Setting/HomePageSectionManager.vue
