+++ src/components/Modal/Setting/SidebarHideManager.vue
